Refactor FollowerList component
In this commit, the <FollowerList /> component was refactored in such a way that it
renders a static data object passed into it as props

Also the user.js was renamed to UserDetails.js while Follower.js was renamed to FollowerDetails.js
